[gen] create soft.generator.common library where the Generator code is implemented
