+++ generators/soft.generator.common/.gitignore
@@ -1,0 +1,2 @@
+/target/
+/.target/
