+++ generators/soft.generator.common/.settings/org.eclipse.jdt.core.prefs
