+++ generators/soft.generator.common/.settings/org.eclipse.jdt.ui.prefs
