+++ generators/soft.generator.cpp/pom.xml
@@ -2,140 +2,25 @@
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>soft.generator.cpp</groupId>
+    
+    <parent>
+        <groupId>soft.generators</groupId>
+        <artifactId>soft.generators</artifactId>
+        <version>1.0.0</version>
+    </parent>
+    
     <artifactId>soft.generator.cpp</artifactId>
-    <version>1.0.0</version>
-
-    <properties>
-        <project.target>.target</project.target>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    </properties>
-
-
-    <repositories>
-        <repository>
-            <id>Eclipse Acceleo</id>
-            <url>https://repo.eclipse.org/content/groups/acceleo</url>
-        </repository>
-    </repositories>
-
-    <pluginRepositories>
-        <pluginRepository>
-            <id>Eclipse Acceleo</id>
-            <url>https://repo.eclipse.org/content/groups/acceleo</url>
-        </pluginRepository>
-    </pluginRepositories>
-
+  
     <dependencies>
         <dependency>
-            <groupId>com.google.guava</groupId>
-            <artifactId>guava</artifactId>
-            <version>26.0-jre</version>
-        </dependency>
-        <dependency>
-            <groupId>commons-cli</groupId>
-            <artifactId>commons-cli</artifactId>
-            <version>1.4</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.acceleo</groupId>
-            <artifactId>org.eclipse.acceleo.engine</artifactId>
-            <version>3.7.6-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.acceleo</groupId>
-            <artifactId>org.eclipse.acceleo.common</artifactId>
-            <version>3.7.6-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.acceleo</groupId>
-            <artifactId>org.eclipse.acceleo.model</artifactId>
-            <version>3.7.6-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.acceleo</groupId>
-            <artifactId>org.eclipse.emf.common</artifactId>
-            <version>2.10.1.v20150123-0348</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.acceleo</groupId>
-            <artifactId>org.eclipse.emf.ecore</artifactId>
-            <version>2.10.2.v20150123-0348</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.acceleo</groupId>
-            <artifactId>org.eclipse.emf.ecore.xmi</artifactId>
-            <version>2.10.2.v20150123-0348</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.acceleo</groupId>
-            <artifactId>org.eclipse.ocl</artifactId>
-            <version>3.4.2.v20140725-2242</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.acceleo</groupId>
-            <artifactId>org.eclipse.ocl.ecore</artifactId>
-            <version>3.3.100.v20140610-0641</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.acceleo</groupId>
-            <artifactId>lpg.runtime.java</artifactId>
-            <version>2.0.17.v201004271640</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.acceleo</groupId>
-            <artifactId>org.eclipse.core.runtime</artifactId>
-            <version>3.10.0.v20140318-2214</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.acceleo</groupId>
-            <artifactId>org.eclipse.core.resources</artifactId>
-            <version>3.9.1.v20140825-1431</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.acceleo</groupId>
-            <artifactId>org.eclipse.equinox.common</artifactId>
-            <version>3.6.200.v20130402-1505</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.acceleo</groupId>
-            <artifactId>org.eclipse.equinox.registry</artifactId>
-            <version>3.5.400.v20140428-1507</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.acceleo</groupId>
-            <artifactId>org.eclipse.ocl.common</artifactId>
-            <version>1.2.0.v20140610-0641</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.acceleo</groupId>
-            <artifactId>org.eclipse.core.contenttype</artifactId>
-            <version>3.4.200.v20140207-1251</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.equinox</groupId>
-            <artifactId>org.eclipse.equinox.preferences</artifactId>
-            <version>3.4.1</version>
-        </dependency>
-        <dependency>
-            <groupId>org.osgi</groupId>
-            <artifactId>org.osgi.framework</artifactId>
-            <version>1.9.0</version>
+            <groupId>soft.generators</groupId>
+            <artifactId>soft.generator.common</artifactId>
+            <version>1.0.0</version>
         </dependency>
     </dependencies>
 
     <build>
-        <sourceDirectory>src</sourceDirectory>
-        <directory>${project.target}</directory>
         <plugins>
-            <plugin>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.8.0</version>
-                <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
-                </configuration>
-            </plugin>
             <plugin>
                 <groupId>org.eclipse.acceleo</groupId>
                 <artifactId>org.eclipse.acceleo.maven</artifactId>
