+++ generators/soft.generator.cpp/src/soft/generator/cpp/generate.mtl
@@ -11,7 +11,6 @@
 [import soft::generator::cpp::tests::generateMock/]
 [import soft::generator::cpp::tests::generateCMake/]
 
-
 [template public generateModel(aPackage : EPackage)]
 [comment @main/]
 [aPackage.generateExports()/]
